dr.medicine/flapi/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
fuzzywuzzy==0.18.0
pandas==1.3.5
python-dotenv==1.0.0
selenium==4.11.2
waitress==2.1.2
